# api/gofiber/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "keyword-search-backend API",
        "title": "keyword-search-backend Swagger API Documentation",
        "contact": {
            "name": "miniyus",
            "url": "https://miniyus.github.io",
            "email": "miniyu97@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/api/groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get all group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "get all groups",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ListResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "update group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "group pk",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "creat group",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.UpdateGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ResponseGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "create group",
                "parameters": [
                    {
                        "description": "creat group",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.CreateGroup"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ResponseGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ValidationErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/groups/name/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get group by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "get group by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ResponseGroup"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/groups/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get group by pk",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "get group by pk",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "pk",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ResponseGroup"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Groups"
                ],
                "summary": "delete group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "type": "bool"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_miniyus_gofiber_internal_api_groups.CreateAction": {
            "type": "object",
            "required": [
                "method",
                "resource"
            ],
            "properties": {
                "method": {
                    "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_permission.Method"
                },
                "resource": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_gofiber_internal_api_groups.CreateGroup": {
            "type": "object",
            "required": [
                "name",
                "permissions"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.CreatePermission"
                    }
                }
            }
        },
        "github_com_miniyus_gofiber_internal_api_groups.CreatePermission": {
            "type": "object",
            "required": [
                "actions",
                "name"
            ],
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.CreateAction"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_gofiber_internal_api_groups.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ResponseGroup"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
        "github_com_miniyus_gofiber_internal_api_groups.ResponseAction": {
            "type": "object",
            "properties": {
                "method": {
                    "$ref": "#/definitions/github_com_miniyus_gofiber_internal_core_permission.Method"
                },
                "resource": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_gofiber_internal_api_groups.ResponseGroup": {
            "type": "object",
            "properties": {
                "actions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.ResponseAction"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_gofiber_internal_api_groups.UpdateGroup": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_miniyus_gofiber_internal_api_groups.CreatePermission"
                    }
                }
            }
        },
        "github_com_miniyus_gofiber_internal_core_api_error.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_miniyus_gofiber_internal_core_api_error.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "failed_fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "github_com_miniyus_gofiber_internal_core_permission.Method": {
            "type": "string",
            "enum": [
                "GET",
                "POST",
                "PUT",
                "PATCH",
                "DELETE"
            ],
            "x-enum-varnames": [
                "GET",
                "POST",
                "PUT",
                "PATCH",
                "DELETE"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token type",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}